# docs/api/openapi.v1.yaml
openapi: 3.0.3
info:
  title: AWT Campus Platform API
  version: 1.0.0
  description: |
    OpenAPI contract aligned to the currently implemented backend scope.

    Implemented now:
    - System health endpoint
    - Room REST API
    - Appointment REST API with conflict checks
    - GraphQL gateway bridge endpoint

    Planned in later increments:
    - Full auth/roles enforcement in backend runtime
    - Notifications and availability domain endpoints
servers:
  - url: http://localhost:4000/api/v1
    description: Local REST API
tags:
  - name: System
  - name: Rooms
  - name: Appointments
  - name: GraphQL
paths:
  /health:
    get:
      tags: [System]
      summary: Health check
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                required: [ok, service]
                properties:
                  ok: { type: boolean }
                  service: { type: string }

  /rooms:
    get:
      tags: [Rooms]
      summary: List rooms
      parameters:
        - in: query
          name: location
          schema: { type: string }
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/RoomStatus'
        - in: query
          name: minCapacity
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Rooms list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
    post:
      tags: [Rooms]
      summary: Create room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomRequest'
      responses:
        '201':
          description: Room created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '400':
          $ref: '#/components/responses/BadRequest'

  /rooms/{id}:
    get:
      tags: [Rooms]
      summary: Get room by id
      parameters:
        - $ref: '#/components/parameters/RoomId'
      responses:
        '200':
          description: Room details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Rooms]
      summary: Update room
      parameters:
        - $ref: '#/components/parameters/RoomId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoomRequest'
      responses:
        '200':
          description: Room updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Rooms]
      summary: Delete room
      parameters:
        - $ref: '#/components/parameters/RoomId'
      responses:
        '204':
          description: Room deleted
        '404':
          $ref: '#/components/responses/NotFound'

  /rooms/{id}/status:
    patch:
      tags: [Rooms]
      summary: Update room status
      parameters:
        - $ref: '#/components/parameters/RoomId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  $ref: '#/components/schemas/RoomStatus'
      responses:
        '200':
          description: Room status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /appointments:
    get:
      tags: [Appointments]
      summary: List appointments
      parameters:
        - in: query
          name: roomId
          schema: { type: string }
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/AppointmentStatus'
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Appointments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AppointmentWithRoom'
    post:
      tags: [Appointments]
      summary: Create appointment
      description: Rejects overlapping non-cancelled appointments in the same room.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppointmentRequest'
      responses:
        '201':
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentWithRoom'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /appointments/{id}:
    get:
      tags: [Appointments]
      summary: Get appointment by id
      parameters:
        - $ref: '#/components/parameters/AppointmentId'
      responses:
        '200':
          description: Appointment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentWithRoom'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Appointments]
      summary: Update appointment
      description: Applies the same time conflict checks as create.
      parameters:
        - $ref: '#/components/parameters/AppointmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAppointmentRequest'
      responses:
        '200':
          description: Appointment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentWithRoom'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Appointments]
      summary: Delete appointment
      parameters:
        - $ref: '#/components/parameters/AppointmentId'
      responses:
        '204':
          description: Appointment deleted
        '404':
          $ref: '#/components/responses/NotFound'

  /appointments/{id}/confirm:
    post:
      tags: [Appointments]
      summary: Confirm appointment
      description: Transition allowed only from DRAFT to CONFIRMED.
      parameters:
        - $ref: '#/components/parameters/AppointmentId'
      responses:
        '200':
          description: Appointment confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentWithRoom'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /appointments/{id}/cancel:
    post:
      tags: [Appointments]
      summary: Cancel appointment
      description: Transition allowed from DRAFT or CONFIRMED to CANCELLED.
      parameters:
        - $ref: '#/components/parameters/AppointmentId'
      responses:
        '200':
          description: Appointment cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentWithRoom'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /graphql:
    post:
      tags: [GraphQL]
      summary: GraphQL gateway endpoint
      description: |
        GraphQL bridge endpoint hosted by the same backend.
        Example supported queries:
        - rooms
        - appointments
        - roomsWithNextAppointments
        - freeRooms
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query: { type: string }
                variables: { type: object, additionalProperties: true }
      responses:
        '200':
          description: GraphQL response envelope
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: object, additionalProperties: true }
                  errors:
                    type: array
                    items:
                      type: object
                      additionalProperties: true

components:
  parameters:
    RoomId:
      in: path
      name: id
      required: true
      schema: { type: string }
    AppointmentId:
      in: path
      name: id
      required: true
      schema: { type: string }

  responses:
    BadRequest:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: Resource conflict (for example overlapping appointment)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Access denied for current role or ownership
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

  schemas:
    ApiError:
      type: object
      required: [type, message]
      properties:
        type: { type: string }
        message: { type: string }
        issues:
          type: array
          items:
            type: object
            additionalProperties: true

    RoomStatus:
      type: string
      enum: [AVAILABLE, OCCUPIED, MAINTENANCE]

    AppointmentStatus:
      type: string
      enum: [DRAFT, CONFIRMED, CANCELLED]

    Room:
      type: object
      required: [id, name, location, capacity, equipment, status, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        location: { type: string }
        capacity: { type: integer }
        equipment:
          type: array
          items: { type: string }
        status:
          $ref: '#/components/schemas/RoomStatus'
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    CreateRoomRequest:
      type: object
      required: [name, location, capacity]
      properties:
        name: { type: string, minLength: 2, maxLength: 120 }
        location: { type: string, minLength: 2, maxLength: 160 }
        capacity: { type: integer, minimum: 1 }
        equipment:
          type: array
          items: { type: string }
          default: []
        status:
          $ref: '#/components/schemas/RoomStatus'

    UpdateRoomRequest:
      type: object
      properties:
        name: { type: string, minLength: 2, maxLength: 120 }
        location: { type: string, minLength: 2, maxLength: 160 }
        capacity: { type: integer, minimum: 1 }
        equipment:
          type: array
          items: { type: string }
        status:
          $ref: '#/components/schemas/RoomStatus'

    Appointment:
      type: object
      required: [id, title, roomId, startsAt, endsAt, participants, status, createdAt, updatedAt]
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        roomId: { type: string }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        participants:
          type: array
          items: { type: string, format: email }
        status:
          $ref: '#/components/schemas/AppointmentStatus'
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    AppointmentWithRoom:
      allOf:
        - $ref: '#/components/schemas/Appointment'
        - type: object
          required: [room]
          properties:
            room:
              $ref: '#/components/schemas/Room'

    CreateAppointmentRequest:
      type: object
      required: [title, roomId, startsAt, endsAt]
      properties:
        title: { type: string, minLength: 2, maxLength: 120 }
        description: { type: string, maxLength: 500 }
        roomId: { type: string }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        participants:
          type: array
          items: { type: string, format: email }
          default: []
        status:
          $ref: '#/components/schemas/AppointmentStatus'

    UpdateAppointmentRequest:
      type: object
      properties:
        title: { type: string, minLength: 2, maxLength: 120 }
        description: { type: string, nullable: true, maxLength: 500 }
        roomId: { type: string }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        participants:
          type: array
          items: { type: string, format: email }
        status:
          $ref: '#/components/schemas/AppointmentStatus'